Refactorizo controller, fixes varios
