+++ src/main/java/com/codeoftheweb/salvo/models/Game.java
@@ -61,7 +61,6 @@ public class Game {
     }
 
     public void addScore(Score score){
-        //score.setGame(this);
         scores.add(score);
     }
 }
