+++ src/main/java/com/codeoftheweb/salvo/models/Player.java
@@ -58,7 +58,6 @@ public class Player {
     }
 
     public void addScore(Score score){
-        //score.setPlayer(this);
         scores.add(score);
     }
