+++ src/main/java/com/codeoftheweb/salvo/models/Score.java
@@ -1,7 +1,5 @@
 package com.codeoftheweb.salvo.models;
 
-import com.codeoftheweb.salvo.models.Game;
-import com.codeoftheweb.salvo.models.Player;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
