+++ src/main/java/com/codeoftheweb/salvo/models/Ship.java
@@ -1,6 +1,5 @@
 package com.codeoftheweb.salvo.models;
 
-import com.codeoftheweb.salvo.models.GamePlayer;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
